requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
Deprecated==1.3.1
easyocr==1.7.2
filelock==3.20.0
firebase_admin==7.1.0
Flask==3.1.3
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
libscrc==1.8.1
line-bot-sdk==3.22.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
promptpay==1.1.9
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzbar==0.1.9
qrcode==8.2
requests==2.32.5
rsa==4.9.1
scikit-image==0.26.0
scipy==1.17.0
setuptools==70.2.0
shapely==2.1.2
six==1.17.0
sympy==1.14.0
tifffile==2026.2.20
torch==2.10.0+cpu
torchvision==0.25.0+cpu
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
yarl==1.22.0
